Move model wrappers to separate package
