+++ fabric-model-loading-api-v1/src/client/java/net/fabricmc/fabric/api/client/model/loading/v1/wrapper/WrapperUnbakedGroupedBlockStateModel.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package net.fabricmc.fabric.api.client.model.loading.v1;
+package net.fabricmc.fabric.api.client.model.loading.v1.wrapper;
 
 import net.minecraft.block.BlockState;
 import net.minecraft.client.render.model.Baker;
